re-style dialogs - some renaming
