+++ app/src/main/java/dev/leonlatsch/photok/ui/components/BaseDialogFragment.kt
@@ -14,33 +14,31 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.viewphoto
+package dev.leonlatsch.photok.ui.components
 
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.View
+import androidx.core.content.res.ResourcesCompat
+import androidx.fragment.app.DialogFragment
 import dev.leonlatsch.photok.R
-import dev.leonlatsch.photok.databinding.DialogBottomSheetDetailsBinding
-import dev.leonlatsch.photok.model.database.entity.Photo
-import dev.leonlatsch.photok.ui.components.BindableBottomSheetDialogFragment
 
 /**
- * Bottom Sheet Dialog for the photo details.
+ * Base for all Dialog Fragments.
+ * Sets drawables for rounded corners.
  *
  * @since 1.0.0
  * @author Leon Latsch
  */
-class DetailsBottomSheetDialog(
-    val photo: Photo?,
-    val photoSize: Int
-) : BindableBottomSheetDialogFragment<DialogBottomSheetDetailsBinding>(R.layout.dialog_bottom_sheet_details) {
+abstract class BaseDialogFragment : DialogFragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        photo ?: dismiss()
-    }
 
-    override fun bind(binding: DialogBottomSheetDetailsBinding) {
-        super.bind(binding)
-        binding.context = this
+        // Set necessary drawables for rounded corners
+        dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+        view.background =
+            ResourcesCompat.getDrawable(resources, R.drawable.bg_dialog_round_light, null)
     }
 }
