+++ app/src/main/res/layout/dialog_bottom_sheet_details.xml
