+++ app/src/main/res/layout/dialog_bottom_sheet_process.xml
@@ -40,7 +40,7 @@
             android:layout_marginStart="10dp"
             android:layout_marginEnd="10dp"
             android:layout_marginBottom="10dp"
-            android:background="@drawable/bg_bottom_sheet_light"
+            android:background="@drawable/bg_dialog_round_light"
             android:orientation="vertical"
             android:paddingStart="10dp"
             android:paddingTop="20dp"
