+++ app/src/main/res/layout/dialog_change_password.xml
@@ -29,10 +29,9 @@
     </data>
 
     <LinearLayout
-        android:layout_width="400dp"
+        android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:animateLayoutChanges="true"
-        android:background="@drawable/bg_bottom_sheet_light"
         android:gravity="center_horizontal"
         android:minWidth="400dp"
         android:orientation="vertical"
